Fully deallocate VM when shutting down
Simply doing `shutdown -h now` doesn't deallocate the virtual machine,
and so doesn't halt billing for the resource.

This commit ensures that the machine is fully shut down and deallocated.
To do this, we add a system-assigned managed identity to the virtual
machine, and give that identity the "Virtual Machine Contributor" role
for the VM.

That allows our `mailbackup-checkshutdown` script to fetch an auth token
from the instance metadata service and call the ARM API to deallocate
the VM.