Shutdown anyway after 30m
It seems like the Azure control plane regularly gets stuck halfway through a VM deallocation. The VM's connection to the instance metadata service is interrupted, meaning we can no longer fetch tokens, but the VM itself never gets shut down. Work around this problem by just shutting down the VM after half an hour.

The risk here is that this is what happens every time, and we end up getting billed for a 100% duty cycle, but I can't see a better way to "unstick" the stuck deallocation at the moment.